refactor: update import paths to use store for Types and add dark mode functionality, this with the aim of creating a global state store for isDarckMode to use in different places in the project
