+++ src/components/ButtonDark.vue
@@ -1,18 +1,5 @@
 <script  setup lang="ts">
-import { ref, type Ref } from 'vue';
-
-let isDarkMode: Ref<boolean> = ref(true);
-const toggleDarkMode = (mode?: string | null) => {
-  isDarkMode.value = mode ? true : !isDarkMode.value;
-  if (isDarkMode.value) {
-    document.documentElement.classList.add('dark');
-    localStorage.setItem('mode', 'dark');
-  } else {
-    document.documentElement.classList.remove('dark');
-    localStorage.setItem('mode', '');
-  }
-}
-toggleDarkMode(localStorage.getItem('mode'));
+import { isDarkMode, toggleDarkMode } from '../store/darkMode';
 </script>
 
 <template>
