+++ src/components/TopBar.vue
@@ -1,10 +1,11 @@
 <script lang="ts" setup>
 import { PopoverGroup } from '@headlessui/vue';
 import Dropdown from './Dropdown.vue';
-import type { HeaderContent } from '../types/Types';
+import type { HeaderContent } from '../store/Types';
 import { useI18n } from 'vue-i18n';
 import { savedLocale } from '../i18n/i18n';
-const { t, getLocaleMessage, locale } = useI18n();
+import { isDarkMode } from '../store/darkMode';
+const { getLocaleMessage, locale } = useI18n();
 
 
 const changeLanguage = (lang: string) => {
@@ -22,8 +23,8 @@ const headerContent: HeaderContent = <HeaderContent>localeMessages.header;
     <nav class="mx-auto h-full flex max-w-7xl items-center justify-between p-2 lg:px-10" aria-label="Global">
       <div class="flex lg:flex-1">
         <a  href= "#" class="-m-1.5 p-1.5 flex gap-2">
-          <img class="h-8 w-auto" src="https://tailwindui.com/plus/img/logos/mark.svg?color=indigo&shade=600" alt="" />
-          <span class="dark:text-white">{{ t('header.logoName') }}</span>
+          <img v-if="!isDarkMode" class="h-8 w-auto" src="https://coderthemes.com/yum/assets/logo-dark-6dbab3e1.png" alt="" />
+          <img v-else class="h-8 w-auto" src="https://coderthemes.com/yum/assets/logo-light-35c89c2c.png" alt="" />
         </a>
       </div>
       <PopoverGroup class="hidden lg:flex lg:gap-x-12">
